Successfull boot with the nvalloc
This introduces a new `CONFIG_NVALLOC` option that enables
the new nvalloc and disables the buddy allocator.

Only the core alloc and free functions of the buddy allocator
have been replaced.
The remainder is either not called or non-functional.

Also the zone->lock is still taken by the free functions,
even if the nvalloc does not need it.